requirements.txt
Flask
Pillow
requests
beautifulsoup4
qrcode
lunar-python